safari-534.53-branch/LayoutTests/media/media-can-play-mpeg-audio-expected.txt

Test HTMLMediaElement canPlayType() method with multiple mp3 MIME types.

EXPECTED (video.canPlayType('audio/x-mpeg') == 'maybe') OK
EXPECTED (video.canPlayType('audio/x-mp3') == 'maybe') OK
EXPECTED (video.canPlayType('audio/mp3') == 'maybe') OK
EXPECTED (video.canPlayType('audio/mp4') == 'maybe') OK
EXPECTED (video.canPlayType('audio/aac') == 'maybe') OK
END OF TEST
